+++ DiscordRoleComparer/Model/DiscordFacade.cs
@@ -1,9 +1,6 @@
 ﻿using Discord;
 using Discord.WebSocket;
-using Newtonsoft.Json;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -11,10 +8,7 @@ namespace DiscordRoleComparer
 {
     public class DiscordFacade
     {
-        public DiscordFacade()
-        {
-            
-        }
+        public DiscordFacade() { }
 
         private List<GuildData> guildDatas = new List<GuildData>();
 
@@ -25,7 +19,7 @@ namespace DiscordRoleComparer
 
         private async Task MainAsync(string token)
         {
-            client = new DiscordSocketClient(new DiscordSocketConfig() { GatewayIntents =  GatewayIntents.AllUnprivileged | GatewayIntents.GuildMembers });
+            client = new DiscordSocketClient(new DiscordSocketConfig() { GatewayIntents = GatewayIntents.AllUnprivileged | GatewayIntents.GuildMembers });
             client.GuildAvailable += OnGuildAvailable;
             await client.LoginAsync(TokenType.Bot, token);
             await client.StartAsync();
@@ -68,39 +62,6 @@ namespace DiscordRoleComparer
 
             guildDatas.Add(guildData);
             return Task.CompletedTask;
-        }
-        #endregion
-
-        #region Save/Load
-        public void OpenJsonFile()
-        {
-            var fileInfo = AskForJsonFile();
-            GuildData guildData = LoadGuildJson(fileInfo);
-            if (guildData == null) return;
-
-            // Debugging Output Below This Line
-            foreach (DiscordMember member in guildData.Members)
-            {
-                Debug.WriteLine(member.ToString());
-            }
-            Debug.WriteLine($"Server Name: {guildData.Name}, Number of Members: {guildData.Members?.Count}, Number of Roles: {guildData.Roles.Count}");
-        }
-
-        private FileInfo AskForJsonFile()
-        {
-            var openFileDialog = new System.Windows.Forms.OpenFileDialog();
-            if (openFileDialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
-            {
-                return new FileInfo(openFileDialog.FileName);
-            }
-            return null;
-        }
-
-        private GuildData LoadGuildJson(FileInfo jsonFile)
-        {
-            if (jsonFile == null) return null;
-            string jsonString = new StreamReader(jsonFile.FullName).ReadToEnd();
-            return JsonConvert.DeserializeObject<GuildData>(jsonString);
         }
         #endregion
     }
